AUC estimator proposed by Chambless and Diao

Chambless and Diao's estimator of cumulative/dynamic AUC for right-censored time-to-event data

Usage

AUC.cd(Surv.rsp, Surv.rsp.new = NULL, lp, lpnew, times)

Arguments

Surv.rsp

A Surv(.,.) object containing to the outcome of the training data.

Surv.rsp.new

A Surv(.,.) object containing the outcome of the test data.

lp

The vector of predictors estimated from the training data.

lpnew

The vector of predictors obtained from the test data.

times

A vector of time points at which to evaluate AUC.

Details

This function implements the estimator of cumulative/dynamic AUC proposed in Section 3.3 of Chambless and Diao (2006). In contrast to the general form of Chambless and Diao's estimator, AUC.cd is restricted to Cox regression. Specifically, it is assumed that lp and lpnew are the predictors of a Cox proportional hazards model. Estimates obtained from AUC.cd are valid as long as the Cox model is specified correctly. The iauc summary measure is given by the integral of AUC on [0, max(times)] (weighted by the estimated probability density of the time-to-event outcome).

Note that the recursive estimators proposed in Sections 3.1 and 3.2 of Chambless and Diao (2006) are not implemented in the survAUC package.

Value

AUC.cd returns an object of class survAUC. Specifically, AUC.cd returns a list with the following components:

auc

The cumulative/dynamic AUC estimates (evaluated at times).

times

The vector of time points at which AUC is evaluated.

iauc

The summary measure of AUC.

References

Chambless, L. E. and G. Diao (2006).
Estimation of time-dependent area under the ROC curve for long-term risk prediction.
Statistics in Medicine 25, 3474–3486.

See also

Examples

data(cancer,package="survival")
TR <- ovarian[1:16,]
TE <- ovarian[17:26,]
train.fit  <- survival::coxph(survival::Surv(futime, fustat) ~ age,
                    x=TRUE, y=TRUE, method="breslow", data=TR)

lp <- predict(train.fit)
lpnew <- predict(train.fit, newdata=TE)
Surv.rsp <- survival::Surv(TR$futime, TR$fustat)
Surv.rsp.new <- survival::Surv(TE$futime, TE$fustat)
times <- seq(10, 1000, 10)                  

AUC_CD <- AUC.cd(Surv.rsp, Surv.rsp.new, lp, lpnew, times)