Gonen and Heller's Concordance Index for Cox models

Gonen and Heller's Concordance Index for Cox proportional hazards models

Usage

GHCI(lpnew)

Arguments

lpnew

The vector of predictors obtained from the test data.

Value

A length-one numeric vector containing the concordance probability estimate.

Details

This function implements the concordance probability estimator proposed by Gonen and Heller (2005). It has the same interpretation as Harrell's C for survival data (implemented in the rcorr.cens function of the Hmisc package).

The results obtained from GHCI are valid as long as lpnew is the predictor of a correctly specified Cox proportional hazards model. In this case, the estimator remains valid even if the censoring times depend on the values of the predictor.

Note that the smoothed version of GHCI, which is proposed in Section 3 of Gonen and Heller (2005), is not implemented in R package survAUC.

References

Harrell, F. E., R. M. Califf, D. B. Pryor, K. L. Lee and R. A. Rosati (1982).
Evaluating the yield of medical tests.
Journal of the American Medical Association 247, 2543–2546.

Harrell, F. E., K. L. Lee, R. M. Califf, D. B. Pryor and R. A. Rosati (1984).
Regression modeling strategies for improved prognostic prediction.
Statistics in Medicine 3, 143–152.

Gonen, M. and G. Heller (2005).
Concordance probability and discriminatory power in proportional hazards regression.
Biometrika 92, 965–970.

See also

Examples

data(cancer,package="survival")
TR <- ovarian[1:16,]
TE <- ovarian[17:26,]
train.fit  <- survival::coxph(survival::Surv(futime, fustat) ~ age,
                    x=TRUE, y=TRUE, method="breslow", data=TR)

lpnew <- predict(train.fit, newdata=TE)
                 
GHCI(lpnew)
#> [1] 0.8515446