Distance-based estimators of survival predictive accuracy

Inverse-probability-of-censoring weighted estimators of absolute and squared deviations between survival functions

Usage

predErr(
  Surv.rsp,
  Surv.rsp.new,
  lp,
  lpnew,
  times,
  type = "brier",
  int.type = "unweighted"
)

Arguments

Surv.rsp

A Surv(.,.) object containing to the outcome of the training data.

Surv.rsp.new

A Surv(.,.) object containing the outcome of the test data.

lp

The vector of predictors estimated from the training data.

lpnew

The vector of predictors obtained from the test data.

times

A vector of time points at which to evaluate the prediction error curve.

type

A string specifying the type of prediction error curve: 'brier' refers to the squared deviation between predicted and observed survival (Brier score), 'robust' refers to the absolute deviation between predicted and observed survival.

int.type

A string specifying the type of integration method for the prediction error curves. Either 'unweighted' or 'weighted'.

Value

predErr returns an object of class survErr. Specifically, predErr returns a list containing the following components:

error

The prediction error estimates (evaluated at times).

times

The vector of time points at which prediction errors are evaluated.

ierror

The integrated prediction error.

Details

This function implements two types of prediction error curves for right-censored time-to-event data: The Brier Score (type = "brier", Gerds and Schumacher 2006) estimates the squared deviation between predicted and observed survival whereas the method proposed by Schmid et al. (2011) estimates the absolute deviation between predicted and observed survival (type = "robust").

Both methods are based on inverse-probability-of-censoring weights and do not assume a specific working model for survival prediction. Note, however, that the estimators implemented in predErr, are restricted to situations where the random censoring assumption holds.

Time-independent summary measures of prediction error are given by the the areas under the prediction error curves. If int.type = "weighted", prediction errors are weighted by the estimated probability density of the time-to-event outcome.

References

Gerds, T. A. and M. Schumacher (2006).
Consistent estimation of the expected Brier score in general survival models with right-censored event times.
Biometrical Journal 48, 1029–1040.

Schmid, M., T. Hielscher, T. Augustin, and O. Gefeller (2011).
A robust alter- native to the Schemper-Henderson estimator of prediction error.
Biometrics 67, 524–535.

See also

Examples

data(cancer,package="survival")
TR <- ovarian[1:16,]
TE <- ovarian[17:26,]
train.fit  <- survival::coxph(survival::Surv(futime, fustat) ~ age, x=TRUE, y=TRUE, 
                    method="breslow", data=TR)

lp <- predict(train.fit)
lpnew <- predict(train.fit, newdata=TE)
Surv.rsp <- survival::Surv(TR$futime, TR$fustat)
Surv.rsp.new <- survival::Surv(TE$futime, TE$fustat)
times <- 1:500                  

predErr(Surv.rsp, Surv.rsp.new, lp, lpnew, times, 
        type = "brier", int.type = "unweighted")
#> $error
#>   [1] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>   [7] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [13] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [19] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [25] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [31] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [37] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [43] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [49] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [55] 0.00000000 0.00000000 0.00000000 0.00000000 0.01203563 0.01203563
#>  [61] 0.01203563 0.01203563 0.01203563 0.01203563 0.01203563 0.01203563
#>  [67] 0.01203563 0.01203563 0.01203563 0.01203563 0.01203563 0.01203563
#>  [73] 0.01203563 0.01203563 0.01203563 0.01203563 0.01203563 0.01203563
#>  [79] 0.01203563 0.01203563 0.01203563 0.01203563 0.01203563 0.01203563
#>  [85] 0.01203563 0.01203563 0.01203563 0.01203563 0.01203563 0.01203563
#>  [91] 0.01203563 0.01203563 0.01203563 0.01203563 0.01203563 0.01203563
#>  [97] 0.01203563 0.01203563 0.01203563 0.01203563 0.01203563 0.01203563
#> [103] 0.01203563 0.01203563 0.01203563 0.01203563 0.01203563 0.01203563
#> [109] 0.01203563 0.01203563 0.01203563 0.01203563 0.01203563 0.01203563
#> [115] 0.04289229 0.04289229 0.04289229 0.04289229 0.04289229 0.04289229
#> [121] 0.04289229 0.04289229 0.04289229 0.04289229 0.04289229 0.04289229
#> [127] 0.04289229 0.04289229 0.04289229 0.04289229 0.04289229 0.04289229
#> [133] 0.04289229 0.04289229 0.04289229 0.04289229 0.04289229 0.04289229
#> [139] 0.04289229 0.04289229 0.04289229 0.04289229 0.04289229 0.04289229
#> [145] 0.04289229 0.04289229 0.04289229 0.04289229 0.04289229 0.04289229
#> [151] 0.04289229 0.04289229 0.04289229 0.04289229 0.04289229 0.09978915
#> [157] 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915
#> [163] 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915
#> [169] 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915
#> [175] 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915
#> [181] 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915
#> [187] 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915
#> [193] 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915
#> [199] 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915
#> [205] 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915
#> [211] 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915
#> [217] 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915
#> [223] 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915
#> [229] 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915
#> [235] 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915
#> [241] 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915
#> [247] 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915
#> [253] 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915
#> [259] 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915
#> [265] 0.09978915 0.09978915 0.09978915 0.02070790 0.02070790 0.02070790
#> [271] 0.02070790 0.02070790 0.02070790 0.02070790 0.02070790 0.02070790
#> [277] 0.02070790 0.02070790 0.02070790 0.02070790 0.02070790 0.02070790
#> [283] 0.02070790 0.02070790 0.02070790 0.02070790 0.02070790 0.02070790
#> [289] 0.02070790 0.02070790 0.02070790 0.02070790 0.02070790 0.02070790
#> [295] 0.02070790 0.02070790 0.02070790 0.02070790 0.02070790 0.02070790
#> [301] 0.02070790 0.02070790 0.02070790 0.02070790 0.02070790 0.02070790
#> [307] 0.02070790 0.02070790 0.02070790 0.02070790 0.02070790 0.02070790
#> [313] 0.02070790 0.02070790 0.02070790 0.02070790 0.02070790 0.02070790
#> [319] 0.02070790 0.02070790 0.02070790 0.02070790 0.02070790 0.02070790
#> [325] 0.02070790 0.02070790 0.02070790 0.02070790 0.11913397 0.11913397
#> [331] 0.11913397 0.11913397 0.11913397 0.11913397 0.11913397 0.11913397
#> [337] 0.11913397 0.11913397 0.11913397 0.11913397 0.11913397 0.11913397
#> [343] 0.11913397 0.11913397 0.11913397 0.11913397 0.11913397 0.11913397
#> [349] 0.11913397 0.11913397 0.11913397 0.11913397 0.16801500 0.16801500
#> [355] 0.16801500 0.16801500 0.16801500 0.16801500 0.16801500 0.16801500
#> [361] 0.16801500 0.16801500 0.16801500 0.16801500 0.20660097 0.20660097
#> [367] 0.20660097 0.20660097 0.20660097 0.20660097 0.20660097 0.20660097
#> [373] 0.20660097 0.20660097 0.20660097 0.20660097 0.20528588 0.20528588
#> [379] 0.20528588 0.20528588 0.20528588 0.20528588 0.20528588 0.20528588
#> [385] 0.20528588 0.20528588 0.20528588 0.20528588 0.20528588 0.20528588
#> [391] 0.20528588 0.20528588 0.20528588 0.20528588 0.20528588 0.20528588
#> [397] 0.20528588 0.20528588 0.20528588 0.20528588 0.20528588 0.20528588
#> [403] 0.20528588 0.20528588 0.20528588 0.20528588 0.20528588 0.20528588
#> [409] 0.20528588 0.20528588 0.20528588 0.20528588 0.20528588 0.20528588
#> [415] 0.20528588 0.20528588 0.20528588 0.20528588 0.20528588 0.20528588
#> [421] 0.20548010 0.20548010 0.20548010 0.20548010 0.20548010 0.20548010
#> [427] 0.20548010 0.20548010 0.20548010 0.20548010 0.16410140 0.16410140
#> [433] 0.16410140 0.16410140 0.16410140 0.16410140 0.16410140 0.16410140
#> [439] 0.16410140 0.16410140 0.16410140 0.16410140 0.16410140 0.16410140
#> [445] 0.16410140 0.16410140 0.16410140 0.16487286 0.16487286 0.16487286
#> [451] 0.16487286 0.16487286 0.16487286 0.16487286 0.16487286 0.16487286
#> [457] 0.16487286 0.16487286 0.16487286 0.16487286 0.16487286 0.16487286
#> [463] 0.16487286 0.14485504 0.14485504 0.14485504 0.14485504 0.14485504
#> [469] 0.14485504 0.14485504 0.14485504 0.14485504 0.14485504 0.14485504
#> [475] 0.13876067 0.13876067 0.14279134 0.14279134 0.14279134 0.14279134
#> [481] 0.14279134 0.14279134 0.14279134 0.14279134 0.14279134 0.14279134
#> [487] 0.14279134 0.14279134 0.14279134 0.14279134 0.14279134 0.14279134
#> [493] 0.14279134 0.14279134 0.14279134 0.14279134 0.14279134 0.14279134
#> [499] 0.14279134 0.14279134
#> 
#> $times
#>   [1]   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18
#>  [19]  19  20  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36
#>  [37]  37  38  39  40  41  42  43  44  45  46  47  48  49  50  51  52  53  54
#>  [55]  55  56  57  58  59  60  61  62  63  64  65  66  67  68  69  70  71  72
#>  [73]  73  74  75  76  77  78  79  80  81  82  83  84  85  86  87  88  89  90
#>  [91]  91  92  93  94  95  96  97  98  99 100 101 102 103 104 105 106 107 108
#> [109] 109 110 111 112 113 114 115 116 117 118 119 120 121 122 123 124 125 126
#> [127] 127 128 129 130 131 132 133 134 135 136 137 138 139 140 141 142 143 144
#> [145] 145 146 147 148 149 150 151 152 153 154 155 156 157 158 159 160 161 162
#> [163] 163 164 165 166 167 168 169 170 171 172 173 174 175 176 177 178 179 180
#> [181] 181 182 183 184 185 186 187 188 189 190 191 192 193 194 195 196 197 198
#> [199] 199 200 201 202 203 204 205 206 207 208 209 210 211 212 213 214 215 216
#> [217] 217 218 219 220 221 222 223 224 225 226 227 228 229 230 231 232 233 234
#> [235] 235 236 237 238 239 240 241 242 243 244 245 246 247 248 249 250 251 252
#> [253] 253 254 255 256 257 258 259 260 261 262 263 264 265 266 267 268 269 270
#> [271] 271 272 273 274 275 276 277 278 279 280 281 282 283 284 285 286 287 288
#> [289] 289 290 291 292 293 294 295 296 297 298 299 300 301 302 303 304 305 306
#> [307] 307 308 309 310 311 312 313 314 315 316 317 318 319 320 321 322 323 324
#> [325] 325 326 327 328 329 330 331 332 333 334 335 336 337 338 339 340 341 342
#> [343] 343 344 345 346 347 348 349 350 351 352 353 354 355 356 357 358 359 360
#> [361] 361 362 363 364 365 366 367 368 369 370 371 372 373 374 375 376 377 378
#> [379] 379 380 381 382 383 384 385 386 387 388 389 390 391 392 393 394 395 396
#> [397] 397 398 399 400 401 402 403 404 405 406 407 408 409 410 411 412 413 414
#> [415] 415 416 417 418 419 420 421 422 423 424 425 426 427 428 429 430 431 432
#> [433] 433 434 435 436 437 438 439 440 441 442 443 444 445 446 447 448 449 450
#> [451] 451 452 453 454 455 456 457 458 459 460 461 462 463 464 465 466 467 468
#> [469] 469 470 471 472 473 474 475 476 477 478 479 480 481 482 483 484 485 486
#> [487] 487 488 489 490 491 492 493 494 495 496 497 498 499 500
#> 
#> $ierror
#> [1] 0.2613864
#> 
#> attr(,"class")
#> [1] "survErr"

predErr(Surv.rsp, Surv.rsp.new, lp, lpnew, times, 
        type = "robust", int.type = "unweighted")
#> $error
#>   [1] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>   [7] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [13] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [19] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [25] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [31] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [37] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [43] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [49] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [55] 0.00000000 0.00000000 0.00000000 0.00000000 0.05157477 0.05157477
#>  [61] 0.05157477 0.05157477 0.05157477 0.05157477 0.05157477 0.05157477
#>  [67] 0.05157477 0.05157477 0.05157477 0.05157477 0.05157477 0.05157477
#>  [73] 0.05157477 0.05157477 0.05157477 0.05157477 0.05157477 0.05157477
#>  [79] 0.05157477 0.05157477 0.05157477 0.05157477 0.05157477 0.05157477
#>  [85] 0.05157477 0.05157477 0.05157477 0.05157477 0.05157477 0.05157477
#>  [91] 0.05157477 0.05157477 0.05157477 0.05157477 0.05157477 0.05157477
#>  [97] 0.05157477 0.05157477 0.05157477 0.05157477 0.05157477 0.05157477
#> [103] 0.05157477 0.05157477 0.05157477 0.05157477 0.05157477 0.05157477
#> [109] 0.05157477 0.05157477 0.05157477 0.05157477 0.05157477 0.05157477
#> [115] 0.10384285 0.10384285 0.10384285 0.10384285 0.10384285 0.10384285
#> [121] 0.10384285 0.10384285 0.10384285 0.10384285 0.10384285 0.10384285
#> [127] 0.10384285 0.10384285 0.10384285 0.10384285 0.10384285 0.10384285
#> [133] 0.10384285 0.10384285 0.10384285 0.10384285 0.10384285 0.10384285
#> [139] 0.10384285 0.10384285 0.10384285 0.10384285 0.10384285 0.10384285
#> [145] 0.10384285 0.10384285 0.10384285 0.10384285 0.10384285 0.10384285
#> [151] 0.10384285 0.10384285 0.10384285 0.10384285 0.10384285 0.17998522
#> [157] 0.17998522 0.17998522 0.17998522 0.17998522 0.17998522 0.17998522
#> [163] 0.17998522 0.17998522 0.17998522 0.17998522 0.17998522 0.17998522
#> [169] 0.17998522 0.17998522 0.17998522 0.17998522 0.17998522 0.17998522
#> [175] 0.17998522 0.17998522 0.17998522 0.17998522 0.17998522 0.17998522
#> [181] 0.17998522 0.17998522 0.17998522 0.17998522 0.17998522 0.17998522
#> [187] 0.17998522 0.17998522 0.17998522 0.17998522 0.17998522 0.17998522
#> [193] 0.17998522 0.17998522 0.17998522 0.17998522 0.17998522 0.17998522
#> [199] 0.17998522 0.17998522 0.17998522 0.17998522 0.17998522 0.17998522
#> [205] 0.17998522 0.17998522 0.17998522 0.17998522 0.17998522 0.17998522
#> [211] 0.17998522 0.17998522 0.17998522 0.17998522 0.17998522 0.17998522
#> [217] 0.17998522 0.17998522 0.17998522 0.17998522 0.17998522 0.17998522
#> [223] 0.17998522 0.17998522 0.17998522 0.17998522 0.17998522 0.17998522
#> [229] 0.17998522 0.17998522 0.17998522 0.17998522 0.17998522 0.17998522
#> [235] 0.17998522 0.17998522 0.17998522 0.17998522 0.17998522 0.17998522
#> [241] 0.17998522 0.17998522 0.17998522 0.17998522 0.17998522 0.17998522
#> [247] 0.17998522 0.17998522 0.17998522 0.17998522 0.17998522 0.17998522
#> [253] 0.17998522 0.17998522 0.17998522 0.17998522 0.17998522 0.17998522
#> [259] 0.17998522 0.17998522 0.17998522 0.17998522 0.17998522 0.17998522
#> [265] 0.17998522 0.17998522 0.17998522 0.10090397 0.10090397 0.10090397
#> [271] 0.10090397 0.10090397 0.10090397 0.10090397 0.10090397 0.10090397
#> [277] 0.10090397 0.10090397 0.10090397 0.10090397 0.10090397 0.10090397
#> [283] 0.10090397 0.10090397 0.10090397 0.10090397 0.10090397 0.10090397
#> [289] 0.10090397 0.10090397 0.10090397 0.10090397 0.10090397 0.10090397
#> [295] 0.10090397 0.10090397 0.10090397 0.10090397 0.10090397 0.10090397
#> [301] 0.10090397 0.10090397 0.10090397 0.10090397 0.10090397 0.10090397
#> [307] 0.10090397 0.10090397 0.10090397 0.10090397 0.10090397 0.10090397
#> [313] 0.10090397 0.10090397 0.10090397 0.10090397 0.10090397 0.10090397
#> [319] 0.10090397 0.10090397 0.10090397 0.10090397 0.10090397 0.10090397
#> [325] 0.10090397 0.10090397 0.10090397 0.10090397 0.19933004 0.19933004
#> [331] 0.19933004 0.19933004 0.19933004 0.19933004 0.19933004 0.19933004
#> [337] 0.19933004 0.19933004 0.19933004 0.19933004 0.19933004 0.19933004
#> [343] 0.19933004 0.19933004 0.19933004 0.19933004 0.19933004 0.19933004
#> [349] 0.19933004 0.19933004 0.19933004 0.19933004 0.24821107 0.24821107
#> [355] 0.24821107 0.24821107 0.24821107 0.24821107 0.24821107 0.24821107
#> [361] 0.24821107 0.24821107 0.24821107 0.24821107 0.28679705 0.28679705
#> [367] 0.28679705 0.28679705 0.28679705 0.28679705 0.28679705 0.28679705
#> [373] 0.28679705 0.28679705 0.28679705 0.28679705 0.27532931 0.27532931
#> [379] 0.27532931 0.27532931 0.27532931 0.27532931 0.27532931 0.27532931
#> [385] 0.27532931 0.27532931 0.27532931 0.27532931 0.27532931 0.27532931
#> [391] 0.27532931 0.27532931 0.27532931 0.27532931 0.27532931 0.27532931
#> [397] 0.27532931 0.27532931 0.27532931 0.27532931 0.27532931 0.27532931
#> [403] 0.27532931 0.27532931 0.27532931 0.27532931 0.27532931 0.27532931
#> [409] 0.27532931 0.27532931 0.27532931 0.27532931 0.27532931 0.27532931
#> [415] 0.27532931 0.27532931 0.27532931 0.27532931 0.27532931 0.27532931
#> [421] 0.27715626 0.27715626 0.27715626 0.27715626 0.27715626 0.27715626
#> [427] 0.27715626 0.27715626 0.27715626 0.27715626 0.25065660 0.25065660
#> [433] 0.25065660 0.25065660 0.25065660 0.25065660 0.25065660 0.25065660
#> [439] 0.25065660 0.25065660 0.25065660 0.25065660 0.25065660 0.25065660
#> [445] 0.25065660 0.25065660 0.25065660 0.25485271 0.25485271 0.25485271
#> [451] 0.25485271 0.25485271 0.25485271 0.25485271 0.25485271 0.25485271
#> [457] 0.25485271 0.25485271 0.25485271 0.25485271 0.25485271 0.25485271
#> [463] 0.25485271 0.24398575 0.24398575 0.24398575 0.24398575 0.24398575
#> [469] 0.24398575 0.24398575 0.24398575 0.24398575 0.24398575 0.24398575
#> [475] 0.24002883 0.24002883 0.25234462 0.25234462 0.25234462 0.25234462
#> [481] 0.25234462 0.25234462 0.25234462 0.25234462 0.25234462 0.25234462
#> [487] 0.25234462 0.25234462 0.25234462 0.25234462 0.25234462 0.25234462
#> [493] 0.25234462 0.25234462 0.25234462 0.25234462 0.25234462 0.25234462
#> [499] 0.25234462 0.25234462
#> 
#> $times
#>   [1]   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18
#>  [19]  19  20  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36
#>  [37]  37  38  39  40  41  42  43  44  45  46  47  48  49  50  51  52  53  54
#>  [55]  55  56  57  58  59  60  61  62  63  64  65  66  67  68  69  70  71  72
#>  [73]  73  74  75  76  77  78  79  80  81  82  83  84  85  86  87  88  89  90
#>  [91]  91  92  93  94  95  96  97  98  99 100 101 102 103 104 105 106 107 108
#> [109] 109 110 111 112 113 114 115 116 117 118 119 120 121 122 123 124 125 126
#> [127] 127 128 129 130 131 132 133 134 135 136 137 138 139 140 141 142 143 144
#> [145] 145 146 147 148 149 150 151 152 153 154 155 156 157 158 159 160 161 162
#> [163] 163 164 165 166 167 168 169 170 171 172 173 174 175 176 177 178 179 180
#> [181] 181 182 183 184 185 186 187 188 189 190 191 192 193 194 195 196 197 198
#> [199] 199 200 201 202 203 204 205 206 207 208 209 210 211 212 213 214 215 216
#> [217] 217 218 219 220 221 222 223 224 225 226 227 228 229 230 231 232 233 234
#> [235] 235 236 237 238 239 240 241 242 243 244 245 246 247 248 249 250 251 252
#> [253] 253 254 255 256 257 258 259 260 261 262 263 264 265 266 267 268 269 270
#> [271] 271 272 273 274 275 276 277 278 279 280 281 282 283 284 285 286 287 288
#> [289] 289 290 291 292 293 294 295 296 297 298 299 300 301 302 303 304 305 306
#> [307] 307 308 309 310 311 312 313 314 315 316 317 318 319 320 321 322 323 324
#> [325] 325 326 327 328 329 330 331 332 333 334 335 336 337 338 339 340 341 342
#> [343] 343 344 345 346 347 348 349 350 351 352 353 354 355 356 357 358 359 360
#> [361] 361 362 363 364 365 366 367 368 369 370 371 372 373 374 375 376 377 378
#> [379] 379 380 381 382 383 384 385 386 387 388 389 390 391 392 393 394 395 396
#> [397] 397 398 399 400 401 402 403 404 405 406 407 408 409 410 411 412 413 414
#> [415] 415 416 417 418 419 420 421 422 423 424 425 426 427 428 429 430 431 432
#> [433] 433 434 435 436 437 438 439 440 441 442 443 444 445 446 447 448 449 450
#> [451] 451 452 453 454 455 456 457 458 459 460 461 462 463 464 465 466 467 468
#> [469] 469 470 471 472 473 474 475 476 477 478 479 480 481 482 483 484 485 486
#> [487] 487 488 489 490 491 492 493 494 495 496 497 498 499 500
#> 
#> $ierror
#> [1] 0.2735577
#> 
#> attr(,"class")
#> [1] "survErr"

predErr(Surv.rsp, Surv.rsp.new, lp, lpnew, times, 
        type = "brier", int.type = "weighted")
#> $error
#>   [1] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>   [7] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [13] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [19] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [25] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [31] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [37] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [43] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [49] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [55] 0.00000000 0.00000000 0.00000000 0.00000000 0.01203563 0.01203563
#>  [61] 0.01203563 0.01203563 0.01203563 0.01203563 0.01203563 0.01203563
#>  [67] 0.01203563 0.01203563 0.01203563 0.01203563 0.01203563 0.01203563
#>  [73] 0.01203563 0.01203563 0.01203563 0.01203563 0.01203563 0.01203563
#>  [79] 0.01203563 0.01203563 0.01203563 0.01203563 0.01203563 0.01203563
#>  [85] 0.01203563 0.01203563 0.01203563 0.01203563 0.01203563 0.01203563
#>  [91] 0.01203563 0.01203563 0.01203563 0.01203563 0.01203563 0.01203563
#>  [97] 0.01203563 0.01203563 0.01203563 0.01203563 0.01203563 0.01203563
#> [103] 0.01203563 0.01203563 0.01203563 0.01203563 0.01203563 0.01203563
#> [109] 0.01203563 0.01203563 0.01203563 0.01203563 0.01203563 0.01203563
#> [115] 0.04289229 0.04289229 0.04289229 0.04289229 0.04289229 0.04289229
#> [121] 0.04289229 0.04289229 0.04289229 0.04289229 0.04289229 0.04289229
#> [127] 0.04289229 0.04289229 0.04289229 0.04289229 0.04289229 0.04289229
#> [133] 0.04289229 0.04289229 0.04289229 0.04289229 0.04289229 0.04289229
#> [139] 0.04289229 0.04289229 0.04289229 0.04289229 0.04289229 0.04289229
#> [145] 0.04289229 0.04289229 0.04289229 0.04289229 0.04289229 0.04289229
#> [151] 0.04289229 0.04289229 0.04289229 0.04289229 0.04289229 0.09978915
#> [157] 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915
#> [163] 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915
#> [169] 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915
#> [175] 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915
#> [181] 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915
#> [187] 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915
#> [193] 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915
#> [199] 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915
#> [205] 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915
#> [211] 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915
#> [217] 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915
#> [223] 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915
#> [229] 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915
#> [235] 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915
#> [241] 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915
#> [247] 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915
#> [253] 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915
#> [259] 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915 0.09978915
#> [265] 0.09978915 0.09978915 0.09978915 0.02070790 0.02070790 0.02070790
#> [271] 0.02070790 0.02070790 0.02070790 0.02070790 0.02070790 0.02070790
#> [277] 0.02070790 0.02070790 0.02070790 0.02070790 0.02070790 0.02070790
#> [283] 0.02070790 0.02070790 0.02070790 0.02070790 0.02070790 0.02070790
#> [289] 0.02070790 0.02070790 0.02070790 0.02070790 0.02070790 0.02070790
#> [295] 0.02070790 0.02070790 0.02070790 0.02070790 0.02070790 0.02070790
#> [301] 0.02070790 0.02070790 0.02070790 0.02070790 0.02070790 0.02070790
#> [307] 0.02070790 0.02070790 0.02070790 0.02070790 0.02070790 0.02070790
#> [313] 0.02070790 0.02070790 0.02070790 0.02070790 0.02070790 0.02070790
#> [319] 0.02070790 0.02070790 0.02070790 0.02070790 0.02070790 0.02070790
#> [325] 0.02070790 0.02070790 0.02070790 0.02070790 0.11913397 0.11913397
#> [331] 0.11913397 0.11913397 0.11913397 0.11913397 0.11913397 0.11913397
#> [337] 0.11913397 0.11913397 0.11913397 0.11913397 0.11913397 0.11913397
#> [343] 0.11913397 0.11913397 0.11913397 0.11913397 0.11913397 0.11913397
#> [349] 0.11913397 0.11913397 0.11913397 0.11913397 0.16801500 0.16801500
#> [355] 0.16801500 0.16801500 0.16801500 0.16801500 0.16801500 0.16801500
#> [361] 0.16801500 0.16801500 0.16801500 0.16801500 0.20660097 0.20660097
#> [367] 0.20660097 0.20660097 0.20660097 0.20660097 0.20660097 0.20660097
#> [373] 0.20660097 0.20660097 0.20660097 0.20660097 0.20528588 0.20528588
#> [379] 0.20528588 0.20528588 0.20528588 0.20528588 0.20528588 0.20528588
#> [385] 0.20528588 0.20528588 0.20528588 0.20528588 0.20528588 0.20528588
#> [391] 0.20528588 0.20528588 0.20528588 0.20528588 0.20528588 0.20528588
#> [397] 0.20528588 0.20528588 0.20528588 0.20528588 0.20528588 0.20528588
#> [403] 0.20528588 0.20528588 0.20528588 0.20528588 0.20528588 0.20528588
#> [409] 0.20528588 0.20528588 0.20528588 0.20528588 0.20528588 0.20528588
#> [415] 0.20528588 0.20528588 0.20528588 0.20528588 0.20528588 0.20528588
#> [421] 0.20548010 0.20548010 0.20548010 0.20548010 0.20548010 0.20548010
#> [427] 0.20548010 0.20548010 0.20548010 0.20548010 0.16410140 0.16410140
#> [433] 0.16410140 0.16410140 0.16410140 0.16410140 0.16410140 0.16410140
#> [439] 0.16410140 0.16410140 0.16410140 0.16410140 0.16410140 0.16410140
#> [445] 0.16410140 0.16410140 0.16410140 0.16487286 0.16487286 0.16487286
#> [451] 0.16487286 0.16487286 0.16487286 0.16487286 0.16487286 0.16487286
#> [457] 0.16487286 0.16487286 0.16487286 0.16487286 0.16487286 0.16487286
#> [463] 0.16487286 0.14485504 0.14485504 0.14485504 0.14485504 0.14485504
#> [469] 0.14485504 0.14485504 0.14485504 0.14485504 0.14485504 0.14485504
#> [475] 0.13876067 0.13876067 0.14279134 0.14279134 0.14279134 0.14279134
#> [481] 0.14279134 0.14279134 0.14279134 0.14279134 0.14279134 0.14279134
#> [487] 0.14279134 0.14279134 0.14279134 0.14279134 0.14279134 0.14279134
#> [493] 0.14279134 0.14279134 0.14279134 0.14279134 0.14279134 0.14279134
#> [499] 0.14279134 0.14279134
#> 
#> $times
#>   [1]   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18
#>  [19]  19  20  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36
#>  [37]  37  38  39  40  41  42  43  44  45  46  47  48  49  50  51  52  53  54
#>  [55]  55  56  57  58  59  60  61  62  63  64  65  66  67  68  69  70  71  72
#>  [73]  73  74  75  76  77  78  79  80  81  82  83  84  85  86  87  88  89  90
#>  [91]  91  92  93  94  95  96  97  98  99 100 101 102 103 104 105 106 107 108
#> [109] 109 110 111 112 113 114 115 116 117 118 119 120 121 122 123 124 125 126
#> [127] 127 128 129 130 131 132 133 134 135 136 137 138 139 140 141 142 143 144
#> [145] 145 146 147 148 149 150 151 152 153 154 155 156 157 158 159 160 161 162
#> [163] 163 164 165 166 167 168 169 170 171 172 173 174 175 176 177 178 179 180
#> [181] 181 182 183 184 185 186 187 188 189 190 191 192 193 194 195 196 197 198
#> [199] 199 200 201 202 203 204 205 206 207 208 209 210 211 212 213 214 215 216
#> [217] 217 218 219 220 221 222 223 224 225 226 227 228 229 230 231 232 233 234
#> [235] 235 236 237 238 239 240 241 242 243 244 245 246 247 248 249 250 251 252
#> [253] 253 254 255 256 257 258 259 260 261 262 263 264 265 266 267 268 269 270
#> [271] 271 272 273 274 275 276 277 278 279 280 281 282 283 284 285 286 287 288
#> [289] 289 290 291 292 293 294 295 296 297 298 299 300 301 302 303 304 305 306
#> [307] 307 308 309 310 311 312 313 314 315 316 317 318 319 320 321 322 323 324
#> [325] 325 326 327 328 329 330 331 332 333 334 335 336 337 338 339 340 341 342
#> [343] 343 344 345 346 347 348 349 350 351 352 353 354 355 356 357 358 359 360
#> [361] 361 362 363 364 365 366 367 368 369 370 371 372 373 374 375 376 377 378
#> [379] 379 380 381 382 383 384 385 386 387 388 389 390 391 392 393 394 395 396
#> [397] 397 398 399 400 401 402 403 404 405 406 407 408 409 410 411 412 413 414
#> [415] 415 416 417 418 419 420 421 422 423 424 425 426 427 428 429 430 431 432
#> [433] 433 434 435 436 437 438 439 440 441 442 443 444 445 446 447 448 449 450
#> [451] 451 452 453 454 455 456 457 458 459 460 461 462 463 464 465 466 467 468
#> [469] 469 470 471 472 473 474 475 476 477 478 479 480 481 482 483 484 485 486
#> [487] 487 488 489 490 491 492 493 494 495 496 497 498 499 500
#> 
#> $ierror
#> [1] 0.1286145
#> 
#> attr(,"class")
#> [1] "survErr"